pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mdeck"
version = "0.1.0"
description = "Convert Markdown to Beamer PDF presentations"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["markdown", "beamer", "latex", "slides", "presentation", "pdf", "pandoc"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Markup",
    "Topic :: Multimedia :: Graphics :: Presentation",
    "Environment :: Console",
]
dependencies = [
    "click>=8.0",
    "python-frontmatter>=1.0",
]

[project.urls]
Homepage = "https://github.com/luiscruz/mdeck"
Repository = "https://github.com/luiscruz/mdeck"
Issues = "https://github.com/luiscruz/mdeck/issues"

[project.scripts]
mdeck = "mdeck.cli:cli"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/mdeck"]
